# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khalaf Olive Groves - Ancient Grove Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List spray / treatment applications
      parameters:
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: treeId
          schema: { type: integer }
        - in: query
          name: treatmentKind
          schema: { type: string }
        - in: query
          name: product
          schema: { type: string }
        - in: query
          name: fromDate
          schema: { type: string }
        - in: query
          name: toDate
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/withholding-watch:
    get:
      operationId: getWithholdingWatch
      tags: [treatments]
      summary: List treatments whose withholding window covers a target harvest date
      parameters:
        - in: query
          name: targetDate
          required: true
          schema: { type: string }
        - in: query
          name: windowDays
          schema: { type: integer, minimum: 1, maximum: 90 }
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: treeId
          schema: { type: integer }
      responses:
        "200":
          description: Active withholding entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WithholdingWatchEntry"
  /treatments/{id}:
    get:
      operationId: getTreatment
      tags: [treatments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Treatment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
    patch:
      operationId: updateTreatment
      tags: [treatments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreatmentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
    delete:
      operationId: deleteTreatment
      tags: [treatments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Irrigation Events ────────────────────────────────────────────────────────
...
  /tree-geometry-records:
    get:
      operationId: listTreeGeometryRecords
      tags: [tree-geometry]
      summary: List tree geometry measurements
      parameters:
        - in: query
          name: treeId
          schema: { type: integer }
        - in: query
          name: groveId
          schema: { type: integer }
        - in: query
          name: fromDate
          schema: { type: string }
        - in: query
          name: toDate
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TreeGeometryRecord"
    post:
      operationId: createTreeGeometryRecord
      tags: [tree-geometry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreeGeometryRecordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeGeometryRecord"
  /tree-geometry-records/{id}:
    get:
      operationId: getTreeGeometryRecord
      tags: [tree-geometry]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeGeometryRecord"
    patch:
      operationId: updateTreeGeometryRecord
      tags: [tree-geometry]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTreeGeometryRecordRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeGeometryRecord"
    delete:
      operationId: deleteTreeGeometryRecord
      tags: [tree-geometry]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Sensors (Phase 4) ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Treatment:
      type: object
      required: [id, workerId, groveId, treatmentKind, product, method, appliedAt, withholdingDays, createdAt]
      properties:
        id: { type: integer }
        workerId: { type: integer }
        workerName: { type: string, nullable: true }
        groveId: { type: integer }
        groveName: { type: string, nullable: true }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        linkedFindId: { type: integer, nullable: true }
        treatmentKind:
          type: string
          enum: [organic_spray, mineral, biological, mechanical, sanitation, foliar_nutrient, other]
        product: { type: string }
        activeIngredient: { type: string, nullable: true }
        rate: { type: number, nullable: true }
        rateUnit: { type: string, nullable: true }
        method:
          type: string
          enum: [foliar, soil, drip, dust, mechanical, other]
        areaHectares: { type: number, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        appliedAt: { type: string, format: date-time }
        withholdingDays: { type: integer }
        weatherConditions: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateTreatmentRequest:
      type: object
      required: [workerId, groveId, treatmentKind, product, method]
      properties:
        workerId: { type: integer }
        groveId: { type: integer }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        linkedFindId: { type: integer, nullable: true }
        treatmentKind:
          type: string
          enum: [organic_spray, mineral, biological, mechanical, sanitation, foliar_nutrient, other]
        product: { type: string }
        activeIngredient: { type: string, nullable: true }
        rate: { type: number, nullable: true }
        rateUnit: { type: string, nullable: true }
        method:
          type: string
          enum: [foliar, soil, drip, dust, mechanical, other]
        areaHectares: { type: number, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        appliedAt: { type: string, format: date-time, nullable: true }
        withholdingDays: { type: integer, nullable: true }
        weatherConditions: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    UpdateTreatmentRequest:
      type: object
      properties:
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        photoIds:
          type: array
          nullable: true
          items: { type: integer }
        linkedFindId: { type: integer, nullable: true }
        treatmentKind:
          type: string
          enum: [organic_spray, mineral, biological, mechanical, sanitation, foliar_nutrient, other]
          nullable: true
        product: { type: string, nullable: true }
        activeIngredient: { type: string, nullable: true }
        rate: { type: number, nullable: true }
        rateUnit: { type: string, nullable: true }
        method:
          type: string
          enum: [foliar, soil, drip, dust, mechanical, other]
          nullable: true
        areaHectares: { type: number, nullable: true }
        treesAffectedCount: { type: integer, nullable: true }
        appliedAt: { type: string, format: date-time, nullable: true }
        withholdingDays: { type: integer, nullable: true }
        weatherConditions: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    WithholdingWatchEntry:
      type: object
      required: [treatmentId, groveId, product, appliedAt, withholdingDays, withholdingEndsAt, daysRemaining]
      properties:
        treatmentId: { type: integer }
        groveId: { type: integer }
        groveName: { type: string, nullable: true }
        treeIds:
          type: array
          nullable: true
          items: { type: integer }
        product: { type: string }
        treatmentKind: { type: string }
        method: { type: string, nullable: true }
        activeIngredient: { type: string, nullable: true }
        appliedAt: { type: string, format: date-time }
        withholdingDays: { type: integer }
        withholdingEndsAt: { type: string, format: date-time }
        daysRemaining: { type: integer }
        targetDate: { type: string }
...
    TreeGeometryRecord:
      type: object
      required: [id, treeId, observedAt, createdAt]
      properties:
        id: { type: integer }
        treeId: { type: integer }
        treeCode: { type: string, nullable: true }
        groveId: { type: integer, nullable: true }
        groveName: { type: string, nullable: true }
        workerId: { type: integer, nullable: true }
        workerName: { type: string, nullable: true }
        observedAt: { type: string, format: date-time }
        trunkDiameterMm: { type: number, nullable: true }
        canopyDiameterM: { type: number, nullable: true }
        treeHeightM: { type: number, nullable: true }
        observedCrownAreaM2: { type: number, nullable: true }
        photoIds: { type: array, items: { type: integer }, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateTreeGeometryRecordRequest:
      type: object
      required: [treeId]
      properties:
        treeId: { type: integer }
        workerId: { type: integer, nullable: true }
        observedAt: { type: string, format: date-time, nullable: true }
        trunkDiameterMm: { type: number, nullable: true }
        canopyDiameterM: { type: number, nullable: true }
        treeHeightM: { type: number, nullable: true }
        observedCrownAreaM2: { type: number, nullable: true }
        photoIds: { type: array, items: { type: integer }, nullable: true }
        notes: { type: string, nullable: true }
    UpdateTreeGeometryRecordRequest:
      type: object
      properties:
        observedAt: { type: string, format: date-time, nullable: true }
        trunkDiameterMm: { type: number, nullable: true }
        canopyDiameterM: { type: number, nullable: true }
        treeHeightM: { type: number, nullable: true }
        observedCrownAreaM2: { type: number, nullable: true }
        photoIds: { type: array, items: { type: integer }, nullable: true }
        notes: { type: string, nullable: true }
...